cs493/hw6-chanall/requirements.txt
Flask==2.2.2
google-cloud-datastore==2.9.0
json2html==1.3.0